Refactor: file acquisition
Pull the attribute value variation-attribute mapping export files
automatically from PlentyMarkets.
Save the corresponding URLs in the config.ini file, which is parsed by
the configparser module.

Simplify the findFile function in openfile.py and add documentation.

Add a example config to the project root and add the config to
gitignore.

Signed-off-by: Sebastian Fricke <sebastian.fricke.linux@gmail.com>