Feature: convert to dataframe & cache
Major update of the application, convert all usage of the csv library to
pandas dataframe handling. Simplify the algorithms and automatically
pull data from plentymarkets. Cache the data from plentymarkets as the
download time of those exports is quite long.

Map the ids for features and properties within the config, aswell as the
URLs for the exports from plentymarkets.

Signed-off-by: Sebastian Fricke <sebastian.fricke.linux@gmail.com>