Major refactor: Shrink down script
Reduce the functionality of the script to properties, features and
texts. The reason for this change is, that my new script:
https://github.com/initBasti/plenty_attribute_export
performs the attribute work, more efficiently. Also, the requirements
that were imposed by the cache(bsddb3) to load the data from
plenty(which is still slow), were cumbersome to deal with.

Signed-off-by: Sebastian Fricke <sebastian.fricke.linux@gmail.com>